\name{expected}
\alias{expected}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Runs Expected Authors from table
}
\description{ Generates Expected Value From the Data Being Tested
}
\usage{
expected(Table, C, N)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Table}{
Table Imported and Being Edited for Lotkas Law
}
  \item{C}{
The value of C generated by LotkasC
}
  \item{N}{
The Value of N Generated by LotkasN
}
}

\value{
The Value Returned is a Expected Authors Table Generated From Authors Table.
}

\author{
Kenneth Buker
}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (Table, C, N) 
{
    value <- Table[,1:1]^N
    nvalue <- 1/value
    part2 <- C*nvalue
    return(part2)
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
