\name{LotkasC}
\alias{LotkasC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Runs the creation of C value in Lotkas Law
}
\description{
Generates Lotkas C value from Equations in Pao(1985)
}
\usage{
LotkasC(N)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{N}{
Value N Generated by LotkasN
}
}

\value{
Value Returned is C.
}

\author{
Kenneth Buker
}


%% ~Make other sections like Warning with \section{Warning }{....} ~


\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (N) 
{
    P <- 20
    increm <- c(1:(P - 1))
    sum <- sum(1/increm^N)
    part1 <- sum
    part2 <- 1/((N - 1) * (P^(N - 1)))
    part3 <- 1/(2 * (P^N))
    part4 <- N/(24 * (P - 1)^(N + 1))
    result <- (part1 + part2 + part3 + part4)
    result <- 1/result
    return(result)
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
