\name{LotkasLaw-package}
\alias{LotkasLaw-package}
\alias{LotkasLaw}
\docType{package}
\title{
LotkasLaw
}
\description{
Runs Lotka's Law which is One of the Special Applications of Zipf's Law in Open Source R.
}
\details{
\tabular{ll}{
Package: \tab LotkasLaw\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-29\cr
License: \tab What license is it under?\cr
}
}
\author{
Kenneth Buker & Dr. Alon Friedman

Maintainer: KCIV4@att.net
}
\references{
http://www.sciencedirect.com/science/article/pii/030645738590055X (Pao 1985)

}
\keyword{ package }